# Removing bounds checks when (de)serializing commands/frames.