--- a/projects/RabbitMQ.Client/client/framing/BasicConsumeOk.cs
+++ b/projects/RabbitMQ.Client/client/framing/BasicConsumeOk.cs
@@ -59,7 +59,7 @@ namespace RabbitMQ.Client.Framing.Impl
 
         public override int WriteArgumentsTo(Span<byte> span)
         {
-            return WireFormatting.WriteShortstr(ref span.GetStart(), _consumerTag, span.Length);
+            return WireFormatting.WriteShortstr(ref span.GetStart(), _consumerTag);
         }
 
         public override int GetRequiredBufferSize()