--- a/projects/RabbitMQ.Client/client/framing/BasicPublish.cs
+++ b/projects/RabbitMQ.Client/client/framing/BasicPublish.cs
@@ -30,6 +30,7 @@
 //---------------------------------------------------------------------------
 
 using System;
+
 using RabbitMQ.Client.client.framing;
 using RabbitMQ.Client.Impl;
 
@@ -66,14 +67,14 @@ namespace RabbitMQ.Client.Framing.Impl
 
         public override ProtocolCommandId ProtocolCommandId => ProtocolCommandId.BasicPublish;
         public override string ProtocolMethodName => "basic.publish";
-        public override bool HasContent => true;
 
         public override int WriteArgumentsTo(Span<byte> span)
         {
-            int offset = WireFormatting.WriteShort(span, default);
-            offset += WireFormatting.WriteShortstr(span.Slice(offset), _exchange);
-            offset += WireFormatting.WriteShortstr(span.Slice(offset), _routingKey);
-            return offset + WireFormatting.WriteBits(span.Slice(offset), _mandatory, _immediate);
+            int length = span.Length;
+            int offset = WireFormatting.WriteShort(ref span.GetStart(), default);
+            offset += WireFormatting.WriteShortstr(ref span.GetOffset(offset), _exchange, length - offset);
+            offset += WireFormatting.WriteShortstr(ref span.GetOffset(offset), _routingKey, length - offset);
+            return offset + WireFormatting.WriteBits(ref span.GetOffset(offset), _mandatory, _immediate);
         }
 
         public override int GetRequiredBufferSize()
@@ -104,14 +105,13 @@ namespace RabbitMQ.Client.Framing.Impl
 
         public override ProtocolCommandId ProtocolCommandId => ProtocolCommandId.BasicPublish;
         public override string ProtocolMethodName => "basic.publish";
-        public override bool HasContent => true;
 
         public override int WriteArgumentsTo(Span<byte> span)
         {
-            int offset = WireFormatting.WriteShort(span, default);
-            offset += WireFormatting.WriteShortstr(span.Slice(offset), _exchange.Span);
-            offset += WireFormatting.WriteShortstr(span.Slice(offset), _routingKey.Span);
-            return offset + WireFormatting.WriteBits(span.Slice(offset), _mandatory, _immediate);
+            int offset = WireFormatting.WriteShort(ref span.GetStart(), default);
+            offset += WireFormatting.WriteShortstr(ref span.GetOffset(offset), _exchange.Span);
+            offset += WireFormatting.WriteShortstr(ref span.GetOffset(offset), _routingKey.Span);
+            return offset + WireFormatting.WriteBits(ref span.GetOffset(offset), _mandatory, _immediate);
         }
 
         public override int GetRequiredBufferSize()