… projects/Benchmarks/Program.cs
… projects/Benchmarks/WireFormatting/DataTypeSerialization.cs
… projects/Benchmarks/WireFormatting/MethodSerialization.cs
… projects/Benchmarks/WireFormatting/PrimitivesSerialization.cs
… projects/RabbitMQ.Client/client/framing/BasicQosOk.cs
… projects/RabbitMQ.Client/client/framing/BasicRecover.cs
… projects/RabbitMQ.Client/client/framing/BasicRecoverAsync.cs
… projects/RabbitMQ.Client/client/framing/BasicRecoverOk.cs
… projects/RabbitMQ.Client/client/framing/BasicReject.cs
… projects/RabbitMQ.Client/client/framing/BasicReturn.cs
… projects/RabbitMQ.Client/client/framing/ChannelClose.cs
… projects/RabbitMQ.Client/client/framing/ChannelCloseOk.cs
… projects/RabbitMQ.Client/client/framing/ChannelFlow.cs
… projects/RabbitMQ.Client/client/framing/ChannelFlowOk.cs
… projects/RabbitMQ.Client/client/framing/ChannelOpen.cs
… projects/RabbitMQ.Client/client/framing/ChannelOpenOk.cs
… projects/RabbitMQ.Client/client/framing/ConfirmSelect.cs
… projects/RabbitMQ.Client/client/framing/ConfirmSelectOk.cs
… projects/RabbitMQ.Client/client/framing/ConnectionBlocked.cs
… projects/RabbitMQ.Client/client/framing/ConnectionClose.cs
… projects/RabbitMQ.Client/client/framing/ConnectionCloseOk.cs
… projects/RabbitMQ.Client/client/framing/ConnectionOpen.cs
… projects/RabbitMQ.Client/client/framing/ConnectionOpenOk.cs
… projects/RabbitMQ.Client/client/framing/ConnectionSecure.cs
… projects/RabbitMQ.Client/client/framing/ConnectionSecureOk.cs
… projects/RabbitMQ.Client/client/framing/ConnectionStart.cs
… projects/RabbitMQ.Client/client/framing/ConnectionStartOk.cs
… projects/RabbitMQ.Client/client/framing/ConnectionTune.cs
… projects/RabbitMQ.Client/client/framing/ConnectionTuneOk.cs
… projects/RabbitMQ.Client/client/framing/ConnectionUnblocked.cs