## Proposed Changes

Simplified version of my previous PR. This is the biggest change. It removes bounds-checking on spans by using `Unsafe.Write` methods in the `NetworkOrderSerializer` classes.

Also split up `WireFormatting` into different files to make it easier to work with (Shared, Read, Write).

## Types of Changes

- [ ] Bug fix (non-breaking change which fixes issue #NNNN)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause an observable behavior change in existing systems)
- [ ] Documentation improvements (corrections, new content, etc)
- [ ] Cosmetic change (whitespace, formatting, etc)

## Checklist

- [X] I have read the `CONTRIBUTING.md` document
- [X] I have signed the CA (see https://cla.pivotal.io/sign/rabbitmq)
- [X] All tests pass locally with my changes
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added necessary documentation (if appropriate)
- [ ] Any dependent changes have been merged and published in related repositories

## Further Comments

Created extensions methods for most cases to make it easier to work with byte offsets rather than having `Span.Slice` calls all over the place. We can allow ourselves to do this since we can guarantee that we have the require amount of bytes to read/write the frames since we request a certain amount of bytes when serializing and read the required amount of bytes and check frame-end markers when deserializing.

Most `span.slice` calls were replace with `ref span.GetOffset` calls references to the spans themselves are replaced with `ref span.GetStart()` calls.

This utilizes ref parameters heavily and results in a hefty speed increase in several cases. A handful of cases turn out slower but this still results in an overall gain in performance and there is more room for improvement with simplifying read/write methods for frames as I made an example of in `BasicAck` and more that can be done in WireFormatting.

### .NET Framework 4.8 improvements:
| Slower                                                                     | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| RabbitMQ.Benchmarks.DataTypeLongStringSerialization.LongstrReadPopulated   |      1.10 |           326.49 |           359.30 |         |
| RabbitMQ.Benchmarks.PrimitivesBool.BoolWrite2(param1: True, param2: False) |      1.09 |             9.97 |            10.87 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.NullWrite                   |      1.04 |            18.04 |            18.74 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableGetSizePopulated       |      1.03 |           260.91 |           268.68 |         |
| RabbitMQ.Benchmarks.MethodChannelClose.ChannelCloseRead                    |      1.02 |            27.19 |            27.77 |         |

| Faster                                                                      | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| --------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| RabbitMQ.Benchmarks.MethodBasicAck.BasicAckWrite                            |      1.63 |            21.22 |            13.05 |         |
| RabbitMQ.Benchmarks.DataTypeArraySerialization.ArrayWriteEmpty              |      1.51 |            21.55 |            14.22 |         |
| RabbitMQ.Benchmarks.DataTypeShortStringSerialization.ShortstrWriteEmpty     |      1.41 |            18.47 |            13.13 |         |
| RabbitMQ.Benchmarks.MethodBasicAck.BasicAckRead                             |      1.25 |            22.14 |            17.69 |         |
| RabbitMQ.Benchmarks.MethodBasicProperties.BasicPropertiesWrite              |      1.25 |            78.22 |            62.60 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableWritePopulated          |      1.25 |           655.88 |           525.35 |         |
| RabbitMQ.Benchmarks.MethodBasicDeliver.BasicPublishWrite                    |      1.22 |            83.87 |            68.81 |         |
| RabbitMQ.Benchmarks.MethodChannelClose.ChannelCloseWrite                    |      1.21 |            33.67 |            27.76 |         |
| RabbitMQ.Benchmarks.DataTypeLongStringSerialization.LongstrGetSizeEmpty     |      1.20 |             1.97 |             1.65 |         |
| RabbitMQ.Benchmarks.DataTypeArraySerialization.ArrayWritePopulated          |      1.17 |           146.53 |           125.56 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.DictGetSize                  |      1.12 |            15.49 |            13.86 |         |
| RabbitMQ.Benchmarks.MethodBasicDeliver.BasicPublishMemoryWrite              |      1.11 |            46.75 |            41.95 |         |
| RabbitMQ.Benchmarks.MethodFramingChannelClose.ChannelCloseWrite             |      1.10 |            70.35 |            63.76 |         |
| RabbitMQ.Benchmarks.MethodFramingBasicPublish.BasicPublishMemoryWrite       |      1.09 |           133.65 |           122.17 |         |
| RabbitMQ.Benchmarks.MethodFramingBasicPublish.BasicPublishWrite             |      1.09 |           200.85 |           183.66 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.DictWrite                    |      1.08 |            44.15 |            40.98 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableGetSizeEmpty            |      1.06 |             8.83 |             8.37 |         |
| RabbitMQ.Benchmarks.DataTypeShortStringSerialization.ShortstrWritePopulated |      1.06 |           145.20 |           137.53 |         |
| RabbitMQ.Benchmarks.PrimitivesLong.LongWrite(value: 12345)                  |      1.05 |            11.42 |            10.83 |         |
| RabbitMQ.Benchmarks.PrimitivesShort.ShortWrite(value: 12345)                |      1.04 |            11.56 |            11.07 |         |
| RabbitMQ.Benchmarks.PrimitivesTimestamp.TimestampWrite                      |      1.04 |            12.25 |            11.75 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.BinaryTableValueWrite        |      1.04 |            43.31 |            41.59 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.IntWrite                     |      1.03 |            22.53 |            21.83 |         |
| RabbitMQ.Benchmarks.PrimitivesLonglong.LonglongWrite(value: 12345)          |      1.03 |            11.96 |            11.60 |         |
| RabbitMQ.Benchmarks.DataTypeLongStringSerialization.LongstrWriteEmpty       |      1.03 |            12.35 |            12.02 |         |
| RabbitMQ.Benchmarks.PrimitivesDecimal.DecimalWrite                          |      1.03 |            13.04 |            12.69 |         |
| RabbitMQ.Benchmarks.DataTypeArraySerialization.ArrayGetSizePopulated        |      1.03 |            43.98 |            42.85 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableWriteEmpty              |      1.03 |            22.55 |            21.98 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableReadPopulated           |      1.02 |           881.56 |           865.89 |         |

### .NET Core 3.1 improvements:
| Slower                                                               | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| RabbitMQ.Benchmarks.DataTypeArraySerialization.ArrayGetSizePopulated |      1.04 |            32.29 |            33.51 |         |
| RabbitMQ.Benchmarks.DataTypeArraySerialization.ArrayReadPopulated    |      1.02 |            79.94 |            81.90 |         |

| Faster                                                                    | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| RabbitMQ.Benchmarks.DataTypeLongStringSerialization.LongstrWriteEmpty     |      4.55 |             8.04 |             1.77 |         |
| RabbitMQ.Benchmarks.MethodBasicAck.BasicAckWrite                          |      1.71 |             3.29 |             1.92 |         |
| RabbitMQ.Benchmarks.MethodBasicAck.BasicAckRead                           |      1.51 |             4.60 |             3.05 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.NullWrite                  |      1.37 |             4.19 |             3.06 |         |
| RabbitMQ.Benchmarks.MethodBasicDeliver.BasicPublishMemoryWrite            |      1.36 |            12.46 |             9.13 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.ArrayWrite                 |      1.33 |            10.29 |             7.72 |         |
| RabbitMQ.Benchmarks.MethodChannelClose.ChannelCloseWrite                  |      1.22 |             7.09 |             5.81 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.IntWrite                   |      1.15 |             5.16 |             4.49 |         |
| RabbitMQ.Benchmarks.MethodBasicDeliver.BasicPublishWrite                  |      1.12 |            32.82 |            29.22 |         |
| RabbitMQ.Benchmarks.DataTypeShortStringSerialization.ShortstrWriteEmpty   |      1.12 |             3.90 |             3.49 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.BinaryTableValueWrite      |      1.08 |            19.38 |            17.90 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.DictWrite                  |      1.07 |            20.33 |            19.00 |         |
| RabbitMQ.Benchmarks.MethodFramingChannelClose.ChannelCloseWrite           |      1.07 |            22.30 |            20.88 |         |
| RabbitMQ.Benchmarks.DataTypeArraySerialization.ArrayWritePopulated        |      1.06 |            68.35 |            64.27 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableWriteEmpty            |      1.06 |            11.40 |            10.76 |         |
| RabbitMQ.Benchmarks.DataTypeTableSerialization.TableWritePopulated        |      1.06 |           328.78 |           310.60 |         |
| RabbitMQ.Benchmarks.MethodBasicProperties.BasicPropertiesWrite            |      1.05 |            35.93 |            34.09 |         |
| RabbitMQ.Benchmarks.MethodFramingBasicPublish.BasicPublishWrite           |      1.05 |            96.07 |            91.71 |         |
| RabbitMQ.Benchmarks.MethodFramingBasicPublish.BasicPublishMemoryWrite     |      1.05 |            53.24 |            50.85 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.DictRead                   |      1.05 |            20.38 |            19.47 |         |
| RabbitMQ.Benchmarks.MethodFramingBasicAck.BasicAckWrite                   |      1.05 |            18.39 |            17.58 |         |
| RabbitMQ.Benchmarks.DataTypeFieldSerialization.BinaryTableValueRead       |      1.04 |            23.88 |            22.88 |         |
| RabbitMQ.Benchmarks.DataTypeLongStringSerialization.LongstrWritePopulated |      1.02 |           228.19 |           223.47 |         |