@stebet 2021-03-16
> Thanks! I will.
> 
> Short question:
> 
> > We can allow ourselves to do this since we can guarantee that we have the require amount of bytes to read/write the frames since we request a certain amount of bytes when serializing and read the required amount of bytes and check frame-end markers when deserializing.
> 
> While I agree on the writing part (client -> server), I'm not sold on the reading part (server -> client).
> If the frame contains erroneous data, we will end up in trouble even if we check the frame end.
> E.g.
> Frame size and end marker are correct, but the short string size is written as 100 but only contains 10 bytes afterwards. => BOOM
> 
> I do fancy performance, but for the receiving side this sounds risky.

Currently it doesn't fail, but it might read garbage data (checks I've done have all bytes return 0, haven't been able to see anything else come up). 

I can add simple checks when reading to make sure we have enough remaining bytes when trying to read frames in the case where we might receive malformed frames from the server. It should still be faster than all the slicing.