@stebet 2021-03-16
After doing some thinking, doing all the checks on the Reads is pretty complicated so I'll change this PR to apply to just writes. That'll simplify the PR a bit and limit it to safer code. I'll revisit the reads later.