@stebet 2021-03-18
@bollhals Reverted the `Read*` methods. Rerunning benchmarks now, will update the original comment when I'm done running them.