@stebet 2021-03-19
> Generally it is looking good, left a few remarks.
> 
> 
> 
> Would you mind sharing the improvements of a simple method? (Disassembly file)

Will do.