@stebet 2021-03-19
Example method improvements from `BasicAck`. Note that the new way is much shorter in ASM instructions as well as being branchless (no jumps, which helps the CPU pipelining).

```assembly
;BasicAck.WriteArgumentsToBefore(System.Span`1<Byte>)
    L0000: push ebp
    L0001: mov ebp, esp
    L0003: push edi
    L0004: push esi
    L0005: push ebx
    L0006: lea eax, [ebp+8]
    L0009: mov edx, [eax]
    L000b: mov eax, [eax+4]
    L000e: lea esi, [ecx+4]
    L0011: mov edi, [esi]
    L0013: mov esi, [esi+4]
    L0016: bswap edi
    L0018: xor ebx, ebx
    L001a: bswap esi
    L001c: add esi, 0
    L001f: adc edi, 0
    L0022: cmp eax, 8
    L0025: jb short L0062
    L0027: mov [edx], esi
    L0029: mov [edx+4], edi
    L002c: cmp dword ptr [ebp+0xc], 8
    L0030: jb short L006d
    L0032: mov eax, [ebp+8]
    L0035: mov edx, [ebp+0xc]
    L0038: sub edx, 8
    L003b: add eax, 8
    L003e: movzx ecx, byte ptr [ecx+0xc]
    L0042: cmp edx, 0
    L0045: jbe short L0073
    L0047: test ecx, ecx
    L0049: jne short L004f
    L004b: xor ecx, ecx
    L004d: jmp short L0054
    L004f: mov ecx, 1
    L0054: mov [eax], cl
    L0056: mov eax, 9
    L005b: pop ebx
    L005c: pop esi
    L005d: pop edi
    L005e: pop ebp
    L005f: ret 8
    L0062: mov ecx, 0x28
    L0067: call System.ThrowHelper.ThrowArgumentOutOfRangeException(System.ExceptionArgument)
    L006c: int3
    L006d: call System.ThrowHelper.ThrowArgumentOutOfRangeException()
    L0072: int3
    L0073: call 0x068175d0
    L0078: int3

;BasicAck.WriteArgumentsToAfter(System.Span`1<Byte>)
    L0000: push edi
    L0001: push esi
    L0002: lea eax, [esp+0xc]
    L0006: mov eax, [eax]
    L0008: lea edx, [ecx+4]
    L000b: mov esi, [edx]
    L000d: mov edx, [edx+4]
    L0010: bswap esi
    L0012: xor edi, edi
    L0014: bswap edx
    L0016: add edx, 0
    L0019: adc esi, 0
    L001c: mov [eax], edx
    L001e: mov [eax+4], esi
    L0021: lea eax, [esp+0xc]
    L0025: mov eax, [eax]
    L0027: movzx edx, byte ptr [ecx+0xc]
    L002b: mov [eax+8], dl
    L002e: mov eax, 9
    L0033: pop esi
    L0034: pop edi
    L0035: ret 8
```